Modifica schemas para serem usados na API e no font-end
