+++ packages/front/package.json
@@ -19,6 +19,7 @@
     "core-js": "^3.8.3",
     "dotenv": "^16.0.0",
     "express": "^4.18.1",
+    "myprisma": "file:../../prisma",
     "vee-validate": "^4.5.11",
     "vue": "^3.2.41",
     "vue-prop-validation-helper": "^3.0.0",
