+++ prisma/src/apiSchemas/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Esquemas de corpo de requisicao para chamadas à API
+ */
+
+export * as users from './users';
+export * as squads from './squads';
+export * as tasks from './tasks';
+export * as voting from './voting';
